% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurDir.R
\name{getCurDir}
\alias{getCurDir}
\title{Print the current file's directory}
\usage{
getCurDir()
}
\description{
Call this function as an addin with cursor in the current file to print the file's directory.
}
