% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCurDir.R
\name{setCurDir}
\alias{setCurDir}
\title{Set the working directory to be the current file's directory}
\usage{
setCurDir()
}
\description{
Call this function as an addin with cursor in the current file to change the working directory.
}
